conda_requirements.txt
# packages in environment at E:\RAMP-EXTERNAL\IP-02\OSTRTA\venv\lazy_graph:
#
# Name                    Version                   Build  Channel
ca-certificates           2019.11.28           hecc5488_0    conda-forge
conda-package-handling    1.6.0            py38h62dcd97_0  
intel-openmp              2019.5                      281    anaconda
matplotlib-base           3.1.2            py38h2981e6d_1    conda-forge
mkl-service               2.3.0            py38hb782905_0    anaconda
numpy-base                1.17.4           py38hc3f5095_0    anaconda
pyqt5-sip                 4.19.18                  pypi_0    pypi
python-dateutil           2.8.1                      py_0    anaconda
python-graphviz           0.13.2                   pypi_0    pypi
